change: methods to add dishes to list    bug fix autotag    refinements - adding robustness for listcollector.

+++ src/main/java/com/meg/atable/service/tag/impl/StandardTagChangeListener.java
@@ -1,23 +1,19 @@
 package com.meg.atable.service.tag.impl;
 
-import com.meg.atable.api.model.TagFilterType;
 import com.meg.atable.api.model.TagType;
-import com.meg.atable.data.entity.DishEntity;
 import com.meg.atable.data.entity.TagEntity;
 import com.meg.atable.data.entity.TagSearchGroupEntity;
-import com.meg.atable.data.repository.DishRepository;
 import com.meg.atable.data.repository.TagRepository;
 import com.meg.atable.data.repository.TagSearchGroupRepository;
 import com.meg.atable.service.tag.TagChangeListener;
 import com.meg.atable.service.tag.TagService;
 import com.meg.atable.service.tag.TagStructureService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.PostConstruct;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.stream.Collectors;
 
 /**
@@ -71,25 +67,25 @@ public class StandardTagChangeListener implements TagChangeListener {
             List<TagEntity> childrenTags = tagStructureService.getDescendantTags(childTag, true);
             childrenTags.add(childTag);
             // remove tags (and children) from oldParent
-            if (origParentTag.getSearchSelect())  {
+            if (origParentTag != null && origParentTag.getSearchSelect()) {
                 // get parent tags
-                List<TagEntity> origParentTags = tagStructureService.getAscendantTags(origParentTag,true );
+                List<TagEntity> origParentTags = tagStructureService.getAscendantTags(origParentTag, true);
                 origParentTags.add(origParentTag);
-                List<Long> groupTagIds = origParentTags.stream().map(t -> t.getId()).collect(Collectors.toList());
-                List<Long> memberTagIds = childrenTags.stream().map(t -> t.getId()).collect(Collectors.toList());
+                List<Long> groupTagIds = origParentTags.stream().map(TagEntity::getId).collect(Collectors.toList());
+                List<Long> memberTagIds = childrenTags.stream().map(TagEntity::getId).collect(Collectors.toList());
                 // delete tag groups with groups in the parenttag, and children in the children tag
-                tagStructureService.deleteTagGroupsByGroupAndMember(groupTagIds,memberTagIds);
+                tagStructureService.deleteTagGroupsByGroupAndMember(groupTagIds, memberTagIds);
             }
 
             // add tags to newParent - if new Parent is SearchSelect
             if (newParentTag.getSearchSelect()) {
-                List<TagEntity> newParentTags = tagStructureService.getAscendantTags(newParentTag,true );
+                List<TagEntity> newParentTags = tagStructureService.getAscendantTags(newParentTag, true);
                 newParentTags.add(newParentTag);
                 // add all children tags to a group with this tag as the group
                 List<TagSearchGroupEntity> groupAssignments = new ArrayList<>();
                 // also add all children tags to any parent groups
-                for (TagEntity parentTag: newParentTags) {
-                    groupAssignments.addAll(tagStructureService.buildGroupAssignments(parentTag.getId(),childrenTags));
+                for (TagEntity parentTag : newParentTags) {
+                    groupAssignments.addAll(tagStructureService.buildGroupAssignments(parentTag.getId(), childrenTags));
                 }
                 tagSearchGroupRepository.save(groupAssignments);
             }
@@ -97,28 +93,30 @@ public class StandardTagChangeListener implements TagChangeListener {
 
 
         // save origParentTag, newParentTag, childTag
-        tagRepository.save(origParentTag);
+        if (origParentTag != null) {
+            tagRepository.save(origParentTag);
+        }
         tagRepository.save(newParentTag);
     }
 
     @Override
-    public void onTagUpdate(TagEntity beforeChange,TagEntity updatedTag) {
+    public void onTagUpdate(TagEntity beforeChange, TagEntity updatedTag) {
         if (TagType.Rating.equals(updatedTag.getTagType())) {
             return;
         }
         if (beforeChange.getSearchSelect().equals(updatedTag.getSearchSelect())) {
             return;
         }
-        List<TagEntity> parentTags = tagStructureService.getAscendantTags(updatedTag,true );
+        List<TagEntity> parentTags = tagStructureService.getAscendantTags(updatedTag, true);
         List<TagEntity> childrenTags = tagStructureService.getDescendantTags(updatedTag, true);
         childrenTags.add(updatedTag);
 
         if (!updatedTag.getSearchSelect()) {
-        parentTags.add(updatedTag);
-            List<Long> groupTagIds = parentTags.stream().map(t -> t.getId()).collect(Collectors.toList());
-            List<Long> memberTagIds = childrenTags.stream().map(t -> t.getId()).collect(Collectors.toList());
+            parentTags.add(updatedTag);
+            List<Long> groupTagIds = parentTags.stream().map(TagEntity::getId).collect(Collectors.toList());
+            List<Long> memberTagIds = childrenTags.stream().map(TagEntity::getId).collect(Collectors.toList());
             // delete tag groups with groups in the parenttag, and children in the children tag
-            tagStructureService.deleteTagGroupsByGroupAndMember(groupTagIds,memberTagIds);
+            tagStructureService.deleteTagGroupsByGroupAndMember(groupTagIds, memberTagIds);
             return;
         }
         // no change of parent here - just either adding to groups or subtracting from groups
@@ -126,12 +124,22 @@ public class StandardTagChangeListener implements TagChangeListener {
 
         // this tag is now a search select.
         // add all children tags to a group with this tag as the group
-        List<TagSearchGroupEntity> groupAssignments = tagStructureService.buildGroupAssignments(updatedTag.getId(),childrenTags);
+        List<TagSearchGroupEntity> groupAssignments = tagStructureService.buildGroupAssignments(updatedTag.getId(), childrenTags);
         // also add all children tags to any parent groups
-        for (TagEntity parentTag: parentTags) {
-            groupAssignments.addAll(tagStructureService.buildGroupAssignments(parentTag.getId(),childrenTags));
+        for (TagEntity parentTag : parentTags) {
+            groupAssignments.addAll(tagStructureService.buildGroupAssignments(parentTag.getId(), childrenTags));
         }
         tagSearchGroupRepository.save(groupAssignments);
         return;
+    }
+
+    @Override
+    public void onTagAdd(TagEntity newTag) {
+        // not used in this implementation
+    }
+
+    @Override
+    public void onTagDelete(TagEntity deletedTag) {
+// not used in this implementation
     }
 }
